+++ tests/test_tasks.py
@@ -21,7 +21,7 @@ from pollbot.tasks.buildhub import get_releases
 from pollbot.tasks.product_details import (product_details, ongoing_versions,
                                            devedition_and_beta_in_sync)
 from pollbot.views.utilities import heartbeat
-from pollbot.utils import Status
+from pollbot.utils import Status, yesterday
 
 
 HERE = os.path.dirname(__file__)
@@ -1137,10 +1137,10 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
         assert received["status"] == Status.EXISTS.value
         assert received["message"] == "Build IDs for this release: 20170914024831"
 
-    def _telemetry_mock_release_query(self, body=None):
+    def _telemetry_mock_release_query(self, body=None, *, build_id='20171009192146'):
         url = ("{}/api/queries/search?q="
-               "Uptake+Firefox+RELEASE+57.0+%2820171009192146%29&include_drafts=true")
-        url = url.format(telemetry.TELEMETRY_SERVER)
+               "Uptake+Firefox+RELEASE+57.0+%28{}%29&include_drafts=true")
+        url = url.format(telemetry.TELEMETRY_SERVER, build_id)
         if body is None:
             body = [{
                 "latest_query_data_id": 5678,
@@ -1169,11 +1169,6 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
         url = '{}/api/query_results/5678'.format(telemetry.TELEMETRY_SERVER)
         self.mocked.get(url, status=200, body=json.dumps(body))
 
-    async def test_telemetry_update_uptake_tasks_returns_error_for_previous_nightly(self):
-        received = await telemetry.update_parquet_uptake('firefox', '56.0a1')
-        assert received["status"] == Status.MISSING.value
-        assert received["message"] == "Telemetry update-parquet metrics landed in Firefox Quantum"
-
     async def test_telemetry_update_uptake_tasks_returns_incomplete_for_no_result(self):
         self._mock_buildhub_search()
         self._telemetry_mock_nightly_query([{
@@ -1182,7 +1177,7 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
             "name": "Uptake Firefox NIGHTLY 57.0a1 20170920"
         }])
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
+        received = await telemetry.main_summary_uptake('firefox', '57.0a1')
         assert received["status"] == Status.INCOMPLETE.value
         assert received["message"] == ("Query still processing.")
 
@@ -1193,12 +1188,21 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
             "query_result": {"data": {"rows": []}}
         })
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
+        received = await telemetry.main_summary_uptake('firefox', '57.0a1')
         assert received["status"] == Status.ERROR.value
         assert received["message"] == ("No result found for your query.")
 
-    async def test_telemetry_update_uptake_tasks_returns_incomplete_for_low_nightly_uptake(self):
+    async def test_telemetry_update_uptake_tasks_returns_error_if_not_authenticated(self):
         self._mock_buildhub_search()
+        self._telemetry_mock_nightly_query({'message': 'You are not authenticated'})
+        with pytest.raises(TaskError) as excinfo:
+            await telemetry.main_summary_uptake('firefox', '57.0a1')
+        message = 'STMO: You are not authenticated'
+        assert str(excinfo.value) == message
+
+    async def test_telemetry_update_uptake_tasks_returns_incomplete_for_low_nightly_uptake(self):
+        build_id = "{}192146".format(yesterday(formating='%Y%m%d'))
+        self._mock_buildhub_search(build_id)
         self._telemetry_mock_nightly_query()
         self._telemetry_mock_query_result({
             "query_result": {"data": {"rows": [
@@ -1206,13 +1210,14 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
             ]}}
         })
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
+        received = await telemetry.main_summary_uptake('firefox', '57.0a1')
         assert received["status"] == Status.INCOMPLETE.value
         assert received["message"] == ("Telemetry uptake for version 57.0a1 "
-                                       "(20171009192146) is 45.32%")
+                                       "({}) is 45.32%".format(build_id))
 
     async def test_telemetry_update_uptake_tasks_should_ignore_copied_queries(self):
-        self._mock_buildhub_search()
+        build_id = "{}192146".format(yesterday(formating='%Y%m%d'))
+        self._mock_buildhub_search(build_id)
         self._telemetry_mock_nightly_query([{
             "latest_query_data_id": 123456789,
             "id": 40198,
@@ -1228,13 +1233,14 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
             ]}}
         })
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
+        received = await telemetry.main_summary_uptake('firefox', '57.0a1')
         assert received["status"] == Status.EXISTS.value
         assert received["message"] == ("Telemetry uptake for version 57.0a1 "
-                                       "(20171009192146) is 65.43%")
+                                       "({}) is 65.43%".format(build_id))
 
     async def test_telemetry_update_uptake_tasks_returns_exists_for_high_nightly_uptake(self):
-        self._mock_buildhub_search()
+        build_id = "{}192146".format(yesterday(formating='%Y%m%d'))
+        self._mock_buildhub_search(build_id)
         self._telemetry_mock_nightly_query()
         url = '{}/api/query_results/5678'.format(telemetry.TELEMETRY_SERVER)
         self.mocked.get(url, status=200, body=json.dumps({
@@ -1243,10 +1249,10 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
             ]}}
         }))
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
+        received = await telemetry.main_summary_uptake('firefox', '57.0a1')
         assert received["status"] == Status.EXISTS.value
         assert received["message"] == ("Telemetry uptake for version 57.0a1 "
-                                       "(20171009192146) is 65.43%")
+                                       "({}) is 65.43%".format(build_id))
 
     async def test_telemetry_update_uptake_tasks_returns_missing_for_no_search_query(self):
         self._mock_buildhub_search()
@@ -1254,40 +1260,43 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
         url = '{}/api/query_results/5678'.format(telemetry.TELEMETRY_SERVER)
         self.mocked.get(url, status=404)
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
+        received = await telemetry.main_summary_uptake('firefox', '57.0a1')
         assert received["status"] == Status.MISSING.value
         assert received["message"] == "Query Result 5678 unavailable (HTTP 404)"
 
     async def test_telemetry_update_uptake_tasks_returns_incomplete_for_low_release_uptake(self):
-        self._mock_buildhub_search()
-        self._telemetry_mock_release_query()
+        build_id = "{}192146".format(yesterday(formating='%Y%m%d'))
+        self._mock_buildhub_search(build_id)
+        self._telemetry_mock_release_query(build_id=build_id)
         self._telemetry_mock_query_result({
             "query_result": {"data": {"rows": [
                 {"ratio": 0.4532, "updated": 19074, "total": 42088}
             ]}}
         })
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0')
+        received = await telemetry.main_summary_uptake('firefox', '57.0')
         assert received["status"] == Status.INCOMPLETE.value
-        message = "Telemetry uptake for version 57.0 (20171009192146) is 45.32%"
+        message = "Telemetry uptake for version 57.0 ({}) is 45.32%".format(build_id)
         assert received["message"] == message
 
     async def test_telemetry_update_uptake_tasks_returns_incomplete_for_high_release_uptake(self):
-        self._mock_buildhub_search()
-        self._telemetry_mock_release_query()
+        build_id = "{}192146".format(yesterday(formating='%Y%m%d'))
+        self._mock_buildhub_search(build_id)
+        self._telemetry_mock_release_query(build_id=build_id)
         self._telemetry_mock_query_result({
             "query_result": {"data": {"rows": [
                 {"ratio": 0.65432, "updated": 27236, "total": 42088}
             ]}}
         })
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0')
+        received = await telemetry.main_summary_uptake('firefox', '57.0')
         assert received["status"] == Status.EXISTS.value
-        message = "Telemetry uptake for version 57.0 (20171009192146) is 65.43%"
+        message = "Telemetry uptake for version 57.0 ({}) is 65.43%".format(build_id)
         assert received["message"] == message
 
     async def test_telemetry_update_uptake_creates_the_query_if_not_found_for_nightly(self):
-        self._mock_buildhub_search()
+        build_id = "{}192146".format(yesterday(formating='%Y%m%d'))
+        self._mock_buildhub_search(build_id)
         self._telemetry_mock_nightly_query([])
 
         url = '{}/api/queries'.format(telemetry.TELEMETRY_SERVER)
@@ -1300,10 +1309,10 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
             "id": 5678
         }))
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0a1')
+        received = await telemetry.main_summary_uptake('firefox', '57.0a1')
         assert received["status"] == Status.INCOMPLETE.value
         assert received["message"] == (
-            "Telemetry uptake calculation for version 57.0a1 (20171009192146) is in progress"
+            "Telemetry uptake calculation for version 57.0a1 ({}) is in progress".format(build_id)
         )
 
     async def test_telemetry_update_uptake_creates_the_query_if_not_found_for_release(self):
@@ -1320,7 +1329,7 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
             "id": 5678
         }))
 
-        received = await telemetry.update_parquet_uptake('firefox', '57.0')
+        received = await telemetry.main_summary_uptake('firefox', '57.0')
         assert received["status"] == Status.INCOMPLETE.value
         assert received["message"] == (
             "Telemetry uptake calculation for version 57.0 (20171009192146) is in progress"
@@ -1334,7 +1343,7 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
         self.mocked.post(url, status=403)
 
         with pytest.raises(TaskError) as excinfo:
-            await telemetry.update_parquet_uptake('firefox', '57.0')
+            await telemetry.main_summary_uptake('firefox', '57.0')
         message = 'Unable to create the new query for 57.0 (20171009192146) (HTTP 403)'
         assert str(excinfo.value) == message
 
@@ -1351,6 +1360,6 @@ https://hg.mozilla.org/releases/mozilla-release/rev/3702966a64c80e17d01f613b0a46
         self.mocked.post(url, status=403)
 
         with pytest.raises(TaskError) as excinfo:
-            await telemetry.update_parquet_uptake('firefox', '57.0')
+            await telemetry.main_summary_uptake('firefox', '57.0')
         message = 'Unable to execute the query n°1234 for 57.0 (20171009192146) (HTTP 403)'
         assert str(excinfo.value) == message